+++ R/AzureVM.R
@@ -9,8 +9,6 @@ AzureRMR::build_template_parameters
 
 globalVariables(c("self", "pool"), "AzureVM")
 
-.AzureVM <- new.env()
-
 
 # adding methods to classes in external package must go in .onLoad
 .onLoad <- function(libname, pkgname)
@@ -22,8 +20,3 @@ globalVariables(c("self", "pool"), "AzureVM")
     options(azure_vm_maxpoolsize=10)
 }
 
-.onUnload <- function(libpath)
-{
-    if(exists("pool", envir=.AzureVM))
-        try(parallel::stopCluster(.AzureVM$pool), silent=TRUE)
-}