+++ R/build_json.R
@@ -145,6 +145,37 @@ build_template_parameters.vmss_config <- function(config, name, login_user, size
 
     do.call(add_parameters, config$options$params)
 
+    # add datadisks to params
+    if(!is_empty(config$datadisks))
+    {
+        # fixup datadisk for scaleset
+        for(i in seq_along(config$datadisks))
+        {
+            config$datadisks[[i]]$vm_spec$lun <- i - 1
+            if(config$datadisks[[i]]$vm_spec$createOption == "attach")
+            {
+                config$datadisks[[i]]$vm_spec$createOption <- "empty"
+                config$datadisks[[i]]$vm_spec$diskSizeGB <- config$datadisks[[i]]$res_spec$diskSizeGB
+                config$datadisks[[i]]$vm_spec$storageAccountType <- config$datadisks[[i]]$res_spec$sku
+            }
+            diskname <- config$datadisks[[i]]$vm_spec$name
+            if(!is.null(diskname))
+            {
+                newdiskname <- paste(name, diskname, i, sep="_")
+                config$datadisks[[i]]$res_spec$name <- newdiskname
+                config$datadisks[[i]]$vm_spec$name <- newdiskname
+            }
+        }
+
+        disk_res_spec <- lapply(config$datadisks, `[[`, "res_spec")
+        null <- sapply(disk_res_spec, is.null)
+
+        add_parameters(
+            dataDisks=lapply(config$datadisks, `[[`, "vm_spec"),
+            dataDiskResources=disk_res_spec[!null]
+        )
+    }
+
     jsonlite::prettify(jsonlite::toJSON(params, auto_unbox=TRUE, null="null"))
 }